Share normalizer code across Scalas
